factor out pattern handling into a new module
